requirements-tensorflow.txt
streamlit
scikit-learn==1.5.2
nltk>=3.8.1
svgling
pandas
numpy
ipykernel
scikeras
gensim
tensorboard
matplotlib
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
tensorflow==2.17.1
langchain
langsmith
lxml
faiss-gpu
langchain-chroma
langchain-openai
chromadb
transformers>=4.34.0
sentence-transformers
langchain_huggingface